+++ common/ECS/entity/registry/ARegistry.tpp
@@ -20,7 +20,7 @@ void ARegistry::registerComponent()
 }
 
 template <typename T>
-void ARegistry::addComponent(int entityId, std::shared_ptr<T> component)
+void ARegistry::addComponent(size_t entityId, std::shared_ptr<T> component)
 {
     registerComponent<T>();
     std::string typeName = typeid(T).name();
@@ -32,7 +32,7 @@ void ARegistry::addComponent(int entityId, std::shared_ptr<T> component)
 }
 
 template <typename T>
-std::shared_ptr<T> ARegistry::getComponent(int entityId) const
+std::shared_ptr<T> ARegistry::getComponent(size_t entityId) const
 {
     std::string typeName = typeid(T).name();
     auto it = _components.find(typeName);
@@ -44,7 +44,7 @@ std::shared_ptr<T> ARegistry::getComponent(int entityId) const
 }
 
 template <typename T>
-void ARegistry::removeComponent(int entityId)
+void ARegistry::removeComponent(size_t entityId)
 {
     const char *typeName = typeid(T).name();
     auto it = _components.find(typeName);
@@ -55,7 +55,7 @@ void ARegistry::removeComponent(int entityId)
 }
 
 template <typename T>
-bool ARegistry::hasComponent(int entityId) const
+bool ARegistry::hasComponent(size_t entityId) const
 {
     std::string typeName = typeid(T).name();
     auto it = _components.find(typeName);
